+++ src/styles/learn.scss
@@ -1,7 +1,7 @@
 @import "_vars.scss";
 @import "~bulma/sass/utilities/initial-variables.sass";
 
-.resources-page {
+.learning-center-page {
     .hero {
 
         .back-to-overview {
